remn_v2 now uses conv_reduce_all
Note that the euclidean_sum_dist must be used to obtain the same result as when summing across euclidean distance matrices, as in the previous version of remn_v2. (For other distance metrics, we could define a general function for summing across matric columns.) Also, remn_v2 now yields a matrix with zeros on the diagonal.